[WIP] Retractions preserve homotopy levels.
This commit has the following unsolved propositions:

* The main goal: `IsRetract A B → HasLevel n A → HasLevel n B`
* Lemma 2.4.3 from HoTT `H x ◾ cong g p ≡ cong f p ◾ H y` (where
  `_◾_` is transitivity
* I need an alternate version of the above that I can get with the
  following two helper-lemmas:
  * `p ◾ (q ◾ r) ≡ p ◾ q ◾ r`
  * `p ◾ q ≡ r → q ≡ sym p ◾ r`

I'm concerned that I can maybe not proove the main goal because I'm a
bit confused about the order of the parameters in
`Cubical.Retract.retract`. My understanding is that `retract f g` states
that *`g`* is a retract of *`f`* - which is a bit counter-intuitive to
me. See also previous commit.